refactor: move theme types
