+++ packages/eui-theme-common/src/global_styling/index.ts
@@ -6,6 +6,6 @@
  * Side Public License, v 1.
  */
 
-export * from './types';
+export * from '../services/theme/types';
 export * from './functions';
 export * from './variables';
